aio-web/requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.11
bcrypt==5.0.0
pytz==2025.2
gunicorn==20.1.0
paho-mqtt==2.1.0